Actualización de proyecto

+++ src/main/java/com/upn/reclutamiento/demo/model/PerfilPostulante.java
@@ -8,7 +8,7 @@ public class PerfilPostulante {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     private String nombre;
     private String apellido;
@@ -19,53 +19,52 @@ public class PerfilPostulante {
     @JoinColumn(name = "usuario_id")
     private Usuario usuario;
 
-    // Getters y setters
-    public Integer getId() {
-        return id;
+    // === Getters y Setters ===
+    public Long getId() { 
+        return id; 
     }
 
-    public void setId(Integer id) {
-        this.id = id;
+    public void setId(Long id) { 
+        this.id = id; 
     }
 
-    public String getNombre() {
-        return nombre;
+    public String getNombre() { 
+        return nombre; 
     }
 
-    public void setNombre(String nombre) {
-        this.nombre = nombre;
+    public void setNombre(String nombre) { 
+        this.nombre = nombre; 
     }
 
-    public String getApellido() {
-        return apellido;
+    public String getApellido() { 
+        return apellido; 
     }
 
-    public void setApellido(String apellido) {
-        this.apellido = apellido;
+    public void setApellido(String apellido) { 
+        this.apellido = apellido; 
     }
 
-    public String getDni() {
-        return dni;
+    public String getDni() { 
+        return dni; 
     }
 
-    public void setDni(String dni) {
-        this.dni = dni;
+    public void setDni(String dni) { 
+        this.dni = dni; 
     }
 
-    public String getCv() {
-        return cv;
+    public String getCv() { 
+        return cv; 
     }
 
-    public void setCv(String cv) {
-        this.cv = cv;
+    public void setCv(String cv) { 
+        this.cv = cv; 
     }
 
-    public Usuario getUsuario() {
-        return usuario;
+    public Usuario getUsuario() { 
+        return usuario; 
     }
 
-    public void setUsuario(Usuario usuario) {
-        this.usuario = usuario;
+    public void setUsuario(Usuario usuario) { 
+        this.usuario = usuario; 
     }
 }
-
